+++ src/statistics.jl
@@ -500,67 +500,220 @@ function apply_statistic(stat::Density2DStatistic,
     apply_statistic(ContourStatistic(levels=stat.levels), scales, coord, aes)
 end
 
-
+"""
+    A general statistic for density plots (e.g. KDE plots and violin plots).
+See [`Geom.density`](@ref Gadfly.Geom.density) or [`Geom.violin`](@ref
+Gadfly.Geom.violin) for more details.
+"""
 struct DensityStatistic <: Gadfly.StatisticElement
-    # Number of points sampled
+    """
+    Number of points sampled for estimate. Powers of two yields better
+    performance.
+    """
     n::Int
-    # Bandwidth used for the kernel density estimation
+
+    """
+    Smoothing bandwidth used for the kernel density estimation. This
+    corresponds to the standard deviation of the `kernel`.
+    """
     bw::Real
+
+    """
+    Multiplicative adjustment of the computed optimal bandwidth. This is a
+    relative adjustment, see `bw` to enforce a specific numerical bandwidth.
+    """
+    adjust::Float64
+
+    """
+    Kernel used for density estimation, see `KernelDensity.jl` for more details.
+    Default is the Normal Distribution.
+    """
+    kernel
+
+    """
+    This parameter only applies in the context of multiple densities. If set to
+    `false` (the default), the densities are computed over the full range of
+    data. If `true`, then each density's range will be computed only over the
+    range of data belonging to that group. This option is incompatible with
+    stacked densities since the ranges might not line up any more.
+    """
+    trim::Bool
+
+    """
+    Method for scaling across multiple estimates. If `:area` (default), all
+    density estimates will have the same area under the curve (prior to trimming
+    ). If `:count`, the areas are scaled proportionally to the total number of
+    observations for each density estimate. If `:peak`, then all densities will
+    have the same maximum peak height.
+    """
+    scale::Symbol
+
+    """
+    Control handling of multiple overlapping densities. The `:dodge` option
+    (default) just overlays each density such that they are in front of each
+    other. The `:stack` option places the densities a top of each other. The
+    `:fill` option is similar to `:stack`, but the stacks are all normalized to
+    a constant height of 1.0. This last option is useful for generating
+    conditional density estimates.
+    """
+    position::Symbol
+
+    """
+    Whether the plot is `:horizontal` or `:vertical`
+    """
+    orientation::Symbol
+
+    """
+    Internal flag that is `true` if this density statistic is a violin plot
+    """
+    isviolin::Bool
 end
-DensityStatistic(; n=256, bandwidth=-Inf) = DensityStatistic(n, bandwidth)
 
-input_aesthetics(stat::DensityStatistic) = [:x, :color]
+function input_aesthetics(stat::DensityStatistic)
+    if stat.isviolin
+        return [:x, :y, :color]
+    elseif stat.orientation == :horizontal
+        return [:x, :color]
+    else
+        return [:y, :color]
+    end
+end
+
 output_aesthetics(stat::DensityStatistic) = [:x, :y, :color]
-default_scales(::DensityStatistic) = [Gadfly.Scale.y_continuous()]
 
-"""
-    Stat.density[(; n=256, bandwidth=-Inf)]
+function _find_output_dims(stat::DensityStatistic)
+    output_dims = Union{Symbol, Nothing}[:x, :y]
+    (stat.orientation == :vertical) && reverse!(output_dims)
 
-Estimate the density of `x` at `n` points, and put the result in `x` and `y`.
-Smoothing is controlled by `bandwidth`.  Used by [`Geom.density`](@ref Gadfly.Geom.density).
-"""
-const density = DensityStatistic
+    groupon = [:color]
+    if stat.isviolin
+        reverse!(output_dims)
+        # For violin plots we need an additional dimension for the density data
+        # so we add an additional dimension on the end
+        push!(output_dims, :width)
+        insert!(groupon, 1, output_dims[1])
+    else
+        # For simple density plots there are no categories so we'll insert a
+        # placeholder value into the first dimension
+        insert!(output_dims, 1, nothing)
+    end
+    output_dims, groupon
+end
 
 function apply_statistic(stat::DensityStatistic,
                          scales::Dict{Symbol, Gadfly.ScaleElement},
                          coord::Gadfly.CoordinateElement,
                          aes::Gadfly.Aesthetics)
-    Gadfly.assert_aesthetics_defined("DensityStatistic", aes, :x)
 
-    if aes.color === nothing
-        isa(aes.x[1], Real) || error("Kernel density estimation only works on Real types.")
+    # For all density/violin plots we're computing a new dimension, the density
+    # dimension. We're also overwriting the other dimensions and part of the
+    # trickiness is tracking which dimension refers to what.
+    # Three output dimensions: (1) grouping (2) evaluation points (i.e where
+    # we're evaluating the KDE) (3) density values
+    output_dims, groupon = _find_output_dims(stat)
 
-        x_f64 = collect(Float64, aes.x)
-
-        window = stat.bw <= 0.0 ? KernelDensity.default_bandwidth(x_f64) : stat.bw
-        f = KernelDensity.kde(x_f64, bandwidth=window, npoints=stat.n)
-        aes.x = collect(Float64, f.x)
-        aes.y = f.density
-    else
-        groups = Dict()
-        for (x, c) in zip(aes.x, cycle(aes.color))
-            if !haskey(groups, c)
-                groups[c] = Float64[x]
-            else
-                push!(groups[c], x)
-            end
+    if stat.isviolin
+        xcat, ycat = Scale.iscategorical(scales, :x), Scale.iscategorical(scales, :y)
+        if xcat && ycat
+            error("Either the x or y aesthetics must be Real for kernel density estimation")
+        elseif xcat && stat.orientation == :horizontal
+            error("Horizontal violins require a continuous x axis for kernel density estimation")
+        elseif ycat && stat.orientation == :vertical
+            error("Vertical violins require a continuous y axis for kernel density estimation")
+        elseif !xcat && !ycat # neither x or y is categorical so we'll assume x is meant to be categorical, see #968
+            new_scale = Scale.x_discrete(order=sortperm(unique(aes.x)))
+            Scale.apply_scale(new_scale, [aes], Gadfly.Data(x=aes.x))
+            scales[:x] = new_scale
+            warn(
+            """
+            Both x and y aesthetics are continuous, violin plots require a
+            categorical variable. Transforming x to be categorical.
+            """)
         end
-
-        colors = Array{RGB{Float32}}(0)
-        aes.x = Array{Float64}(0)
-        aes.y = Array{Float64}(0)
-        for (c, xs) in groups
-            window = stat.bw <= 0.0 ? KernelDensity.default_bandwidth(xs) : stat.bw
-            f = KernelDensity.kde(xs, bandwidth=window, npoints=stat.n)
-            append!(aes.x, f.x)
-            append!(aes.y, f.density)
-            for _ in 1:length(f.x)
-                push!(colors, c)
-            end
+        if getfield(aes, output_dims[1]) == nothing
+            setfield!(aes, output_dims[1], fill(1.0, length(getfield(aes, output_dims[2]))))
         end
-        aes.color = discretize_make_ia(colors)
+    elseif getfield(aes, output_dims[2]) == nothing
+        error("The $(output_dims[2]) aesthetic is required for $(stat.orientation) density plots")
     end
-    aes.y_label = Gadfly.Scale.identity_formatter
+
+    grouped_data = Gadfly.groupby(aes, groupon, output_dims[2])
+
+
+    n_pts = stat.position == :fill ? stat.n : stat.n + 2
+    n_groups = length(grouped_data)
+
+    groups = Array{Float64}(n_groups)
+    eval_points = fill(0.0, n_groups, n_pts)
+    densities = fill(0.0, n_groups, n_pts)
+    colors = Array{eltype(aes.color)}(n_groups)
+
+    # if the densities are stacked then we'll need to clamp them so that they
+    # share the same evaluation points (e.g. x values)
+    boundary = extrema(getfield(aes, output_dims[2]))
+
+    for (idx, (keys, belongs)) in enumerate(grouped_data)
+        input = getfield(aes, output_dims[2])[belongs]
+        window = stat.bw <= 0.0 ? KernelDensity.default_bandwidth(input)*stat.adjust : stat.bw
+        (stat.trim) && (boundary = extrema(input))
+        kde_est = KernelDensity.kde(input, kernel=stat.kernel,
+                                     boundary=boundary,
+                                     npoints=stat.n,
+                                     bandwidth=window)
+        evalpts = kde_est.x
+        density = kde_est.density
+        # only store category information if this is a violin plot and we need it
+        if stat.isviolin
+            groups[idx] = keys[1]
+            colors[idx] = keys[2]
+        elseif length(keys) == 1
+            colors[idx] = keys[1]
+        else
+            error("Density plots do not support grouping by more than two dimensions.")
+        end
+        # scale density output depending on `scale` flag
+        scaled_density = stat.position == :fill ? density : vcat(0.0, density, 0.0)
+        if stat.scale == :count
+            scaled_density .*= sum(input)
+        elseif stat.scale == :peak
+            scaled_density ./= maximum(density)
+        end
+
+        eval_points[idx, :] = stat.position == :fill ? evalpts : vcat(boundary[1], evalpts, boundary[2])
+        densities[idx, :] = scaled_density
+    end
+
+    if stat.position == :dodge
+        # if this is a violin plot, make sure to set the grouping
+        (stat.isviolin) && setfield!(aes, output_dims[1], repeat(groups, outer=n_pts))
+        setfield!(aes, output_dims[2], vec(eval_points))
+        setfield!(aes, output_dims[3], vec(densities))
+        (aes.color != nothing) && (aes.color = repeat(colors, outer=n_pts))
+    elseif stat.position == :stack || stat.position == :fill
+        if stat.position == :fill
+            densities ./= sum(densities, 1)
+        end
+        stacked_densities = hcat(copy(densities), fill(0.0, size(densities)...))
+        for i in 1:n_groups
+            for j in 1:i-1
+                stacked_densities[i, 1:n_pts] .+= densities[j, :]
+                stacked_densities[i, n_pts+1:2*n_pts] .+= densities[j, end:-1:1]
+            end
+        end
+        setfield!(aes, output_dims[2], vec(hcat(eval_points, eval_points[:, end:-1:1])))
+        setfield!(aes, output_dims[3], vec(stacked_densities))
+        (aes.color != nothing) && (aes.color = repeat(colors, outer=n_pts*2))
+    end
+
+    if stat.isviolin
+        pad = 0.1
+        maxwidth = maximum(aes.width)
+        broadcast!(*, aes.width, aes.width, 1 - pad)
+        broadcast!(/, aes.width, aes.width, maxwidth)
+    else
+        scales[:y] = Scale.y_continuous()
+        aes.y_label = Gadfly.Scale.identity_formatter
+    end
 end
 
 
@@ -758,7 +911,7 @@ data with `coverage_weight`; and of having a nice numbering with
          granularity_weight=1/4,
          simplicity_weight=1/6,
          coverage_weight=1/3,
-         niceness_weight=1/4) = 
+         niceness_weight=1/4) =
     TickStatistic("x",
               granularity_weight, simplicity_weight, coverage_weight, niceness_weight, ticks)
 
@@ -1620,65 +1773,6 @@ function apply_statistic(stat::QQStatistic,
         Scale.apply_scale(scales[:y], [aes], data)
     end
 end
-
-
-struct ViolinStatistic <: Gadfly.StatisticElement
-    n::Int # Number of points sampled
-end
-ViolinStatistic() = ViolinStatistic(300)
-
-input_aesthetics(::ViolinStatistic) = [:x, :y, :color]
-output_aesthetics(::ViolinStatistic) = [:x, :y, :width, :color]
-default_scales(stat::ViolinStatistic) = [Gadfly.Scale.x_discrete(), Gadfly.Scale.y_continuous()]
-
-### very similar to Stat.density; Geom.violin could be refactored to us it instead
-"""
-    Stat.violin[(n=300)]
-
-Transform $(aes2str(input_aesthetics(violin()))).
-"""
-const violin = ViolinStatistic
-
-function apply_statistic(stat::ViolinStatistic,
-                         scales::Dict{Symbol, Gadfly.ScaleElement},
-                         coord::Gadfly.CoordinateElement,
-                         aes::Gadfly.Aesthetics)
-
-    isa(aes.y[1], Real) || error("Kernel density estimation only works on Real types.")
-
-    grouped_y = Dict(1=>aes.y)
-    grouped_color = Dict{Int, Gadfly.ColorOrNothing}(1=>nothing)
-    ux = unique(aes.x)
-    uxflag = length(ux) < length(aes.x)
-    colorflag = aes.color != nothing
-
-    uxflag && (grouped_y = Dict(x=>aes.y[aes.x.==x] for x in ux))
-
-    grouped_color = (colorflag ? Dict(x=>first(aes.color[aes.x.==x]) for x in ux) : 
-        uxflag && Dict(x=>nothing for x in ux) )
-
-    aes.x     = Array{Float64}(0)
-    aes.y     = Array{Float64}(0)
-    aes.width = Array{Float64}(0)
-    colors = eltype(aes.color)[]
-
-    for (x, ys) in grouped_y
-        window = stat.n > 1 ? KernelDensity.default_bandwidth(ys) : 0.1
-        f = KernelDensity.kde(ys, bandwidth=window, npoints=stat.n)
-        append!(aes.x, fill(x, length(f.x)))
-        append!(aes.y, f.x)
-        append!(aes.width, f.density)
-        append!(colors, fill(grouped_color[x], length(f.x)))
-    end
-
-    colorflag && (aes.color = colors)
-
-    pad = 0.1
-    maxwidth = maximum(aes.width)
-    broadcast!(*, aes.width, aes.width, 1 - pad)
-    broadcast!(/, aes.width, aes.width, maxwidth)
-end
-
 
 struct JitterStatistic <: Gadfly.StatisticElement
     vars::Vector{Symbol}